requirements.txt
accelerate==1.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
efficientnet_pytorch==0.7.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
modelscope==1.31.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psutil==7.1.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
timm==1.0.20
tokenizers==0.21.4
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1

triton==3.4.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xformers==0.0.32.post2
